+++ Editor/Drawers/FindComponentAttributeDrawer.cs
@@ -42,6 +42,10 @@ namespace SaintsField.Editor.Drawers
             if(result != null)
             {
                 onGUIPayload.SetValue(result);
+                if(ExpandableIMGUIScoop.IsInScoop)
+                {
+                    property.serializedObject.ApplyModifiedProperties();
+                }
             }
             return true;
         }