+++ Editor/Drawers/GetPrefabWithComponentAttributeDrawer.cs
@@ -40,6 +40,10 @@ namespace SaintsField.Editor.Drawers
             if(result != null)
             {
                 onGUIPayload.SetValue(result);
+                if(ExpandableIMGUIScoop.IsInScoop)
+                {
+                    property.serializedObject.ApplyModifiedProperties();
+                }
             }
             return true;
         }