+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.2.4 ##
+
+1.  IMGUI: fix `Expandable` won't save the changed value.
+2.  Fix `Required` won't work in parent of an inherited class.
+3.  Fix `FieldType` won't work in array/list.
+4.  Allow `FieldType` with `compType=null` to by-pass the issue that Unity won't show all the prefabs with expected component in the picker.
+
 ## 3.2.3 ##
 
 1.  IMGUI: fix disposed access.